requirements.txt
smolagents>=0.1.0
pygithub>=1.58.0
python-dotenv>=1.0.0
pytest>=7.0.0
black>=23.0.0
tqdm>=4.65.0